(bug)Don't updateCommitIndex if index is greater than commits.length

+++ client/js/utils/fileTreeUtils.js
@@ -2,6 +2,9 @@ var _ = require('underscore');
 
 module.exports.updateTree = function(currentCommit, fileTree, direction) {
   var filepath;
+  if (!currentCommit || !currentCommit.files) {
+    return null;
+  }
   if (direction === 'backward') {
     currentCommit.files.forEach(function(file) {
       file.status === 'added' || file.status === 'renamed' ? removeFile(fileTree, file) : addFile(fileTree, file);
